fetcher/requirements.txt
influxdb-client
tortoise-orm
aiohttp
asyncpg
requests